+++ crates/holochain/src/core/workflow/sys_validation_workflow.rs
@@ -42,6 +42,8 @@ mod chain_test;
 mod test_ideas;
 #[cfg(test)]
 mod tests;
+#[cfg(test)]
+mod validate_op_tests;
 
 #[instrument(skip(
     workspace,
@@ -107,7 +109,7 @@ async fn sys_validation_workflow_inner(
                     let dna_def = DnaDefHashed::from_content_sync((*workspace.dna_def()).clone());
 
                     let r =
-                        validate_op(&op, &dna_def, &cascade, Some(incoming_dht_ops_sender)).await;
+                        validate_op(&op, &dna_def, &cascade, Some(&incoming_dht_ops_sender)).await;
                     r.map(|o| (op_hash, o, dependency))
                 }
                 .boxed()
@@ -190,7 +192,7 @@ pub(crate) async fn validate_op(
     op: &DhtOp,
     dna_def: &DnaDefHashed,
     cascade: &impl Cascade,
-    incoming_dht_ops_sender: Option<IncomingDhtOpSender>,
+    incoming_dht_ops_sender: Option<&impl DhtOpSender>,
 ) -> WorkflowResult<Outcome> {
     match validate_op_inner(op, cascade, dna_def, incoming_dht_ops_sender).await {
         Ok(_) => Ok(Outcome::Accepted),
@@ -241,6 +243,7 @@ fn handle_failed(error: &ValidationOutcome) -> Outcome {
         ValidationOutcome::PrivateEntryLeaked => Rejected,
         ValidationOutcome::PreflightResponseSignature(_) => Rejected,
         ValidationOutcome::UpdateTypeMismatch(_, _) => Rejected,
+        ValidationOutcome::UpdateHashMismatch(_, _) => Rejected,
         ValidationOutcome::VerifySignature(_, _) => Rejected,
         ValidationOutcome::WrongDna(_, _) => Rejected,
         ValidationOutcome::ZomeIndex(_) => Rejected,
@@ -252,7 +255,7 @@ async fn validate_op_inner(
     op: &DhtOp,
     cascade: &impl Cascade,
     dna_def: &DnaDefHashed,
-    incoming_dht_ops_sender: Option<IncomingDhtOpSender>,
+    incoming_dht_ops_sender: Option<&impl DhtOpSender>,
 ) -> SysValidationResult<()> {
     check_entry_visibility(op)?;
     match op {
@@ -302,7 +305,7 @@ async fn validate_op_inner(
                     check_and_hold_store_record(
                         &ActionHash::with_data_sync(&action),
                         cascade,
-                        incoming_dht_ops_sender.clone(),
+                        incoming_dht_ops_sender,
                         dependency_check,
                     )
                     .await?;
@@ -396,7 +399,7 @@ async fn sys_validate_record_inner(
         maybe_entry: Option<&Entry>,
         cascade: &impl Cascade,
     ) -> SysValidationResult<()> {
-        let incoming_dht_ops_sender = None;
+        let incoming_dht_ops_sender: Option<&IncomingDhtOpSender> = None;
         store_record(action, cascade).await?;
         if let Some(maybe_entry) = maybe_entry {
             store_entry(
@@ -455,7 +458,7 @@ async fn register_agent_activity(
     action: &Action,
     cascade: &impl Cascade,
     dna_def: &DnaDefHashed,
-    incoming_dht_ops_sender: Option<IncomingDhtOpSender>,
+    incoming_dht_ops_sender: Option<&impl DhtOpSender>,
 ) -> SysValidationResult<()> {
     // Get data ready to validate
     let prev_action_hash = action.prev_action();
@@ -533,7 +536,7 @@ async fn store_entry(
 async fn register_updated_content(
     entry_update: &Update,
     cascade: &impl Cascade,
-    incoming_dht_ops_sender: Option<IncomingDhtOpSender>,
+    incoming_dht_ops_sender: Option<&impl DhtOpSender>,
 ) -> SysValidationResult<()> {
     // Get data ready to validate
     let original_action_address = &entry_update.original_action_address;
@@ -553,7 +556,7 @@ async fn register_updated_content(
 async fn register_updated_record(
     entry_update: &Update,
     cascade: &impl Cascade,
-    incoming_dht_ops_sender: Option<IncomingDhtOpSender>,
+    incoming_dht_ops_sender: Option<&impl DhtOpSender>,
 ) -> SysValidationResult<()> {
     // Get data ready to validate
     let original_action_address = &entry_update.original_action_address;
@@ -574,7 +577,7 @@ async fn register_updated_record(
 async fn register_deleted_by(
     record_delete: &Delete,
     cascade: &impl Cascade,
-    incoming_dht_ops_sender: Option<IncomingDhtOpSender>,
+    incoming_dht_ops_sender: Option<&impl DhtOpSender>,
 ) -> SysValidationResult<()> {
     // Get data ready to validate
     let removed_action_address = &record_delete.deletes_address;
@@ -596,7 +599,7 @@ async fn register_deleted_by(
 async fn register_deleted_entry_action(
     record_delete: &Delete,
     cascade: &impl Cascade,
-    incoming_dht_ops_sender: Option<IncomingDhtOpSender>,
+    incoming_dht_ops_sender: Option<&impl DhtOpSender>,
 ) -> SysValidationResult<()> {
     // Get data ready to validate
     let removed_action_address = &record_delete.deletes_address;
@@ -618,7 +621,7 @@ async fn register_deleted_entry_action(
 async fn register_add_link(
     link_add: &CreateLink,
     _cascade: &impl Cascade,
-    _incoming_dht_ops_sender: Option<IncomingDhtOpSender>,
+    _incoming_dht_ops_sender: Option<&impl DhtOpSender>,
 ) -> SysValidationResult<()> {
     check_tag_size(&link_add.tag)?;
     Ok(())
@@ -627,7 +630,7 @@ async fn register_add_link(
 async fn register_delete_link(
     link_remove: &DeleteLink,
     cascade: &impl Cascade,
-    incoming_dht_ops_sender: Option<IncomingDhtOpSender>,
+    incoming_dht_ops_sender: Option<&impl DhtOpSender>,
 ) -> SysValidationResult<()> {
     // Get data ready to validate
     let link_add_address = &link_remove.link_add_address;
@@ -642,9 +645,10 @@ async fn register_delete_link(
 
 fn update_check(entry_update: &Update, original_action: &Action) -> SysValidationResult<()> {
     check_new_entry_action(original_action)?;
+    // This shouldn't fail due to the above `check_new_entry_action` check
     let original_action: NewEntryActionRef = original_action
         .try_into()
-        .expect("This can't fail due to the above check_new_entry_action");
+        .map_err(|_| ValidationOutcome::NotNewEntry(original_action.clone()))?;
     check_update_reference(entry_update, &original_action)?;
     Ok(())
 }